Move pkg/forwarder to cmd/traffic/cmd/agent/fwd
This is step 1/2 in the refactor that splits the Interceptor into a
generic Forwarder and an Interceptor.

This commit doesn't change any logic. It just moves the package.
