+++ cmd/traffic/cmd/agent/fwd/http.go
@@ -1,4 +1,4 @@
-package forwarder
+package fwd
 
 import (
 	"bufio"
