+++ cmd/traffic/cmd/agent/fwd/http_test.go
@@ -1,4 +1,4 @@
-package forwarder
+package fwd
 
 import (
 	"net/http"
