+++ cmd/traffic/cmd/agent/fwd/tcp.go
@@ -1,4 +1,4 @@
-package forwarder
+package fwd
 
 import (
 	"context"
